+++ integration/keycloak/helper-keycloak.R
@@ -618,6 +618,7 @@ local_test_options <- function(.local_envir = parent.frame()) {
   withr::local_options(
     list(
       shinyOAuth.skip_browser_token = TRUE,
+      shinyOAuth.allow_insecure_oidc_loopback = TRUE,
       shinyOAuth.timeout = 10
     ),
     .local_envir = .local_envir